airflow/requirements.txt

kaggle
apache-airflow-providers-google
pyarrow
